requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
crispy-bootstrap4==2022.1
Django==4.0.10
django-bootstrap-v5==1.0.11
django-crispy-forms==2.0
django-taggit==4.0.0
django-upgrade==1.14.0
soupsieve==2.4.1
sqlparse==0.4.4
tokenize-rt==5.1.0
typing_extensions==4.7.0
